+++ poem/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# Unreleased
+
+- Add `I18NArgs::set` method.
+
 # [1.2.30] 2022-1-11
 
 - Change the behavior of the `TestClient::query` method to add a KV pair.